Added FETCH_VOICEMAIL intent definition in VoicemailContract.
Content provider does not directly use this intent. But it is fired by
the contacts app to request the voicemail source to fetch voicemail
audio of a single voicemail message.

The constant needs to live in the contracts file to serve as a common
definition between contacts and voicemail source apps.

Bug: 5114261
Change-Id: Ibf7b2a4c871d3561d2ca50d03f8182352fadd3ab

Former-commit-id: 267a2ec0402470f46079ae2d9d5059e2bb67453a