Fix behavior of custom animations for LayoutTransition.
Previously, setting custom animations on a LayoutTransition would cause
those animations to be run not only for changing views in the container, but
for the parent hierarchy of those views as well. This can lead to unexpected
behavior, as seen in the ApiDemos LayoutAnimations and LayoutAnimationsHideShow.
This change changes the behavior so that the parent hierarchy is animated by
the default animations (which change the bounds and scrollX/Y fields) instead
of custom animations.

Change-Id: I9a04d97fabbc34dc0d5809eb3fd8ac08e0801d7c

Former-commit-id: e115ffeb3a05f440c0062ad9b3954b7fefef4b00