Fix stuck in pairing when BT is turned off.
When pairing is stuck, on turning BT off, we were
not setting the outgoing parining variable to null in
setBondState because of the extra check for isEnabled.

isEnabled check was added to prevent the proxy crash in
setProfilePriorities function. Proxies should never be null.
Add a safety check and also some extra logs to debug the problem.

Change-Id: I694dfeb8fa9426b3916775ca868c2313fa9c22ee

Former-commit-id: 89d55ad781c6f1eba77342c89136dce41a7de72a