Fix NumberPicker text positioning.
NumberPicker was making incorrect assumptions about how text was
positioned and centered inside of its controls. This fix changes
the logic of how the scroll wheel draws its text items based on where
the text is positioned inside the EditText that displays the text
by default. This should ensure that the text of the scroll wheel
exactly overlays the text in the EditText view.

Change-Id: I4a7694931f19042bd4142232612490c2ab912d11

Former-commit-id: eeafd4296b7cf4c2ad9c985562f6d70e2a3ffc03