Fix bug 5202226 - Leave menu button visible for no-button phones if targetSdk < ICS
Compatibility support for existing apps. Apps in the wild that
properly target <=GB phones and HC tablets should be able to expect
that their small-screen UI will continue to work as intended on ICS
devices. Make sure that we show the soft menu button in the nav bar if
the current device was not officially supported in Honeycomb, the app
does not target ICS or newer, and the window does not have an action
bar.

Change-Id: I069c30841d3827a60ef8040831fde6f4c051d0ee

Former-commit-id: 82349c54ce793cb0a7cf3acf1bcc6df1e6413abc