Tweaks to NotificationPanel animation
The animation that runs when the NotificationPanel appears
used to start, then pause for a long time as the window/surface/layer
was created, then by the time it started to be visible, the animation
was over. This new approach delays starting the animation until the
layer has been drawn, so the animation can actually run a few frames after
that before finishing.

Change-Id: I998f01fd48cb762178021ad99e2b919b58a1ef3f

Former-commit-id: 41bff38d3060dbcb55133cedaf5d962c3082efc2