Throw an exception if a both nav and system bars are configured
It was possible for a device to be configured to use a navigation bar
(by setting config_showNavigationBar to true in a config.xml file)
and to use a system bar (by being a tablet device of sufficient size).
The display metrics would be set up for both, but the SystemUI would ignore
the navigation bar, leading to conflict in what would be displayed and how
much room there was on the display. This change throws an exception,
causing SystemUI to fail, when this conflict is detected.

Change-Id: Ib5c19d50ff9444bc33fd53d3dcfa016adc4872ef

Former-commit-id: bc0009d48df58305d3330e9716f4ebbca1e36d3d