Add method to retrieve MSISDN for CDMA/LTE devices.
For CDMA/LTE devices, the MDN and MSISDN may be different.
Add a new method getMsisdn() to the Phone interface to return the
MSISDN. For GSM/UMTS, this will be the same as getLine1Number().
For CDMA/LTE, getLine1Number() will continue to return the MDN
and getMsisdn() will return the MSISDN.

Change-Id: Iba0ca24858992b21f63ae7ec0c27d2e90d4b0903

Former-commit-id: 5d39ec4a37de500ada84225f8bd2c244db0cd287