Use the correct dependency
$(LOCAL_BUILT_MODULE) of the module framework points to a
product-specific file (required by dex-preopt), which is nuked by installclean.
While the result of $(java-lib-deps) points to a file in the common
directory.

This change reduces the incremental build time significantly:
It takes around 7 minutes to run "make instalclean && make" without this
change v.s. 3 minutes with this change.
So it will save lots of build time when switching between user, userdebug
and eng builds on the build server.

Change-Id: I832bafca04677af561bb0c28e2e0260f633b96a1

Former-commit-id: b1f57728d9b0122abfc1426052a5a1ab464a934e