Can't stop the fling! Bug: 5335420
Fixed a bug in VelocityTracker where the output velocity was
not being set to zero when not available.

Added a condition to ensure that the velocity is at least
the minimum fling velocity before continuing.  If not, then
the user is trying to stop the fling and scroll more precisely.

Change-Id: I36634b0c3f7a9a09cf20c33f71d41163a8e33eed

Former-commit-id: b0c71eb9f50ce06327fa6bb6219f0970e04fd856