Settings are restored in alphabetical order rather in order of dependency.
Settings were restore in alphabetical order and capturing dependency
among them required keys to be chosen in such a way that after sorting
they apprear in dependency order. Now settings are exported and restored
in the order they are declared in the arrays of settings to backup.
Hence, the order in this array will capture the dependency order.

bug:5343351

Change-Id: I93a40bcdd194943cd6f85aa18f1557d546e38274

Former-commit-id: a571a5836cdbc71feefb13648004b8ca7fe05dfb