Set enable_puk_unlock_screen true by default.
Make pukUnlockScreen the default way to enter PUK ocde.
There are two ways to enter PUK code, pukUnlockScreen UI
and MMI Code for Emergency Dialer. As far as we know most
of carriers are fine with either one though they may prefer
pukUnlockScreen UI. It can be overwrite if specific device or
carrier don't want to show pukUnlockSreen UI.

Note the Emergency Dialer will not show up
if the device is not voice capable. Another reason to make
pukUnlockSreen UI enable as default.

bug:5243771
Change-Id: I141324bef6ab812243a6cbb89870f71c60e838ec

Former-commit-id: 5209864a2bcdc089fe3c433fa7aad87f43dbd200