Fix apn bearer logic.
The original change said that if the RAT were X or Y we would only
accept APN's that explicitly called out X or Y.  This meant that
any device using X or Y would stop working until their APN db were
adjusted.

This change changes it to be if a particular APN calls out X or Y
it will only be considered if the current RAT matches.  If the APN
doesn't specify, it matches all RAT.

This allows just as tight a restriction, but the default is looser.

Change-Id: Ia5e92f13c5052e890bf169e0db9584302afb36f5

Former-commit-id: 5840639fd18c334c0df30ae03b270c452fee0a61