Update all RSSIs whenever the connectivity changes.
The connectivity includes the network condition, which is a
magic number that indicates how useful the connection is
(driven by ability to connect to GTalk, and manifested with
a differently-colored signal icon).

Previously we were only updating the icons of the network
type currently in use. Now we tickle all the icons any time
the connectivity changes, to avoid the two icons getting out
of sync (example: on mobile, connection is bad, connect to
wifi, conncetion is good; wifi icon appears and turns blue
but mobile RSSI remains gray).

Bug: 5369405
Change-Id: I78b62084b819c83f83aacc146feeb323c1199dc0

Former-commit-id: 7c12eee6439435d7f29d49cab76edaedbd809460