Change DCS of MO SMS to the correct value for UCS-2 encoding.
An incorrect value for TP-Data-Coding-Scheme was used for MO SMS with
UCS-2 encoding. According to 3GPP TS 23.038 section 4, when bit 4 is
set to 0, bits 0 and 1 are reserved and have no message class meaning.
We were setting those bits to 1, when they should be set to 0.

This change is required to pass a GCF test case.

Change-Id: If47361f95b0e14accc6220854783dcff3d82abf6

Former-commit-id: df0c2a94b51ead16a8b7c7b3159295bc978d9831