Reworked biometric weak check functions
Renamed isBiometricEnabled to isBiometricWeakInstalled.  This function
now checks if the system property is set, the facelock package is
installed, and if the phone has a front facing camera.  It no longer
checks if facelock is currently set as the unlock method.
Added isBiometricWeakInstalled checks to all cases where facelock is used
in LockPatternKeyguardView

Change-Id: Ia86a7ad6118101c6aab90ffb2ee9c42bf2548149

Former-commit-id: 58396984ef49080d5550919130fc0d869ccf840b