Add a cached formatted phone number to call log.
This column is used by the phone application to cache the formatted
phone number. This is needed to avoid computing the formatted phone
number in the main thread (because it requires disk access) but not
update the formatted phone number after the UI has been shown (except
the first time around or if anything else has changed).

Bug: 5316982
Change-Id: I4e2dd2e3fb35a0e0d15039e4c0c9841a909f7a8a

Former-commit-id: 0fce15b03c06cf3ece42a1ec636349f525f16ea1