Bug 5385279: IOOB in createMisspelledSuggestionSpan
Race condition: when spell check result come back, the spell check span may no longer
be in the text, although it should have been removed by removeSpellCheckSpan.

This is possible because onGetSuggestions is called from an other thread
and it can see the text in an inconsistent state where the span may have been removed
from the text, but the listener that calls removeSpellCheckSpan has not yet been called.
As a result, onGetSuggestions still lists the span but getSpanStart returns -1.

Change-Id: Ia40f7ed2b46fc94888fce5a6b209fb0d0d529bbe

Former-commit-id: c08ec615d26508c14c44680ffe649d46be6de8c5