Don't redraw when not scrolling, even if finger is held down.
Signal the native side that we are not scrolling when a drag is
held motionless to stop screen updates.

Additionally, in the case that a drag is being held motionless,
send a timed message to keep the scroll bars on screen otherwise
they fade in and out, again causing repaints that shouldn't be
necessary.

Bug: 5440335
Change-Id: I94db0323879885fa48924b41d29e8af03e3b5e7f

Former-commit-id: fd8d3dede2f3d1b22e1c69d4c14d392d0aa2b2ab