AAH: Don't panic unless you are sure.
Change the definition of the panic threshold for the time sync service
to take the round trip time of the packet producing the discipline
action into account.  The idea here is that, in the worst case, the
error in the measurement of the clock sync error might be as high as
the round trip time of the packet used to measure the error.  If a
packet says we are 50mSec fast, its RTT is 75mSec, then we could be
anywhere from 25mSec slow to 125mSec fast.

This change basically makes it so that the error measurement must be
greater than the RTT + the fixed panic threshold before we panic.  IOW
we don't panic unless we absolutely sure that our error must be
greater than the fixed panic threshold.  Also, lower the current value
of the panic threshold from 50mSec down to 20.  Now that the
definition takes into account the RTT, we can afford to be much
tighter in our definition of the fixed panic threshold.

Change-Id: I88406203ba4d6ec8ece79c0a09e5f3b28f67c6e8

Former-commit-id: 5b8d7f9de91d59d3f2bf4af07a0822006b56976f