Issue 5432115: Speed up notification scrolling
The NotificationPanel was using views that had non-1 alpha
values set on them (permanently). This is costly in the GL
implementation and caused more rendering overhead, and worse
performance, than simple opaque views would.

The fix is to set the text color and ImageView drawable alpha
directly, without setting the View alpha property.

Change-Id: I381e0bd45bf45784b8e364a27a339e6583189a43

Former-commit-id: ba1fe8e3cfcf029a9bbd9bc7575e2b5c4f4ad8b0