AccessibilityServiceInfo does not report properly the feedback type.
Since the feedback type is an integer the bits of which correspond to a given
feedback type, the AccessibilityNodeInfo#feedbackTypeToString should return a
string representation of all feedback types. The current implementation
incorrectly assumes that the argument had only one flag set. Note that this
is a new ICS API and we would like to have this fixed.

bug:5486933

Change-Id: I5e461c385e199cce4c3054907ff0acf5e9d92afc

Former-commit-id: bb1b9ea8a488eec833b0be18c45dae8505876042