Fix 3384942: Force ASCII-capable IME to show on lockscreen
Only Latin-1 based passwords are supported on the system at
this point.  As such, we need a way to show a Latin-1-compliant
keyboard when unlocking the device.

It is expected that 3rd party IMEs honor this mode for Latin-1
password dialogs. If not, this change will cause lockscreen to
revert to using the Android system IME for entering the password.

Change-Id: I59bb56f6f5968b5abc50cf1b9b26c19ca4f183e4

Former-commit-id: 214406a8e95348bfea8a65d981e508f35bf5cf21